# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * E-Commerce Inventory & Sales System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  productId: number;
  productName: string;
  category: string;
  price: number;
  stockQuantity: number;
  description: string;
  isLowStock: boolean;
}

export interface CreateProductBody {
  productName: string;
  category: string;
  price: number;
  stockQuantity: number;
  description: string;
}

export interface Customer {
  customerId: number;
  name: string;
  email: string;
  phone: string;
  address: string;
}

export interface CreateCustomerBody {
  name: string;
  email: string;
  phone: string;
  address: string;
}

export interface OrderItem {
  orderDetailId: number;
  productId: number;
  productName: string;
  quantity: number;
  subtotal: number;
  price: number;
}

export interface OrderWithDetails {
  orderId: number;
  customerId: number;
  customerName: string;
  orderDate: string;
  totalAmount: number;
  status: string;
  items: OrderItem[];
}

export interface OrderLineInput {
  productId: number;
  quantity: number;
}

export interface CreateOrderBody {
  customerId: number;
  items: OrderLineInput[];
}

export type UpdateOrderStatusBodyStatus =
  (typeof UpdateOrderStatusBodyStatus)[keyof typeof UpdateOrderStatusBodyStatus];

export const UpdateOrderStatusBodyStatus = {
  pending: "pending",
  processing: "processing",
  shipped: "shipped",
  delivered: "delivered",
  cancelled: "cancelled",
} as const;

export interface UpdateOrderStatusBody {
  status: UpdateOrderStatusBodyStatus;
}

export interface InventoryLog {
  logId: number;
  productId: number;
  productName: string;
  stockBefore: number;
  stockAfter: number;
  updatedAt: string;
}

export interface UpdateInventoryBody {
  productId: number;
  newStock: number;
}

export interface DashboardStats {
  totalProducts: number;
  totalOrders: number;
  totalRevenue: number;
  totalCustomers: number;
  lowStockCount: number;
  pendingOrders: number;
}

export interface TopProduct {
  productId: number;
  productName: string;
  category: string;
  totalSold: number;
  totalRevenue: number;
}

export interface CategorySales {
  category: string;
  totalOrders: number;
  totalRevenue: number;
}

export interface SalesTrendPoint {
  date: string;
  totalRevenue: number;
  totalOrders: number;
}

export interface TopCustomer {
  customerId: number;
  name: string;
  email: string;
  totalOrders: number;
  totalSpent: number;
}

export interface InventoryStatus {
  category: string;
  totalStock: number;
  productCount: number;
  lowStockCount: number;
}

export type ListProductsParams = {
  category?: string;
  search?: string;
  lowStock?: boolean;
};

export type ListCustomersParams = {
  search?: string;
};

export type ListOrdersParams = {
  status?: string;
  customerId?: number;
};

export type ListInventoryLogsParams = {
  productId?: number;
};

export type GetTopProductsParams = {
  limit?: number;
};

export type GetSalesTrendParams = {
  days?: number;
};

export type GetTopCustomersParams = {
  limit?: number;
};