# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: E-Commerce Inventory & Sales System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerBody"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{customerId}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer by ID
      parameters:
        - in: path
          name: customerId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
    put:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - in: path
          name: customerId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerBody"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - in: path
          name: customerId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Orders ──────────────────────────────────────────────────────────────────
...
  /analytics/top-products:
    get:
      operationId: getTopProducts
      tags: [analytics]
      summary: Top selling products by revenue
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Top products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopProduct"
...
  /analytics/sales-trend:
    get:
      operationId: getSalesTrend
      tags: [analytics]
      summary: Daily sales totals for trend chart
      parameters:
        - in: query
          name: days
          schema:
            type: integer
      responses:
        "200":
          description: Sales trend data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalesTrendPoint"
  /analytics/top-customers:
    get:
      operationId: getTopCustomers
      tags: [analytics]
      summary: Customers with highest total purchases
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Top customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopCustomer"
...
components:
  schemas:
...
    Customer:
      type: object
      properties:
        customerId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
      required: [customerId, name, email, phone, address]
    CreateCustomerBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
      required: [name, email, phone, address]
...
    TopProduct:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        category:
          type: string
        totalSold:
          type: integer
        totalRevenue:
          type: number
      required: [productId, productName, category, totalSold, totalRevenue]
...
    SalesTrendPoint:
      type: object
      properties:
        date:
          type: string
        totalRevenue:
          type: number
        totalOrders:
          type: integer
      required: [date, totalRevenue, totalOrders]
    TopCustomer:
      type: object
      properties:
        customerId:
          type: integer
        name:
          type: string
        email:
          type: string
        totalOrders:
          type: integer
        totalSpent:
          type: number
      required: [customerId, name, email, totalOrders, totalSpent]
...