# lib/api-zod/src/generated/types/customer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * E-Commerce Inventory & Sales System API
 * OpenAPI spec version: 0.1.0
 */

export interface Customer {
  customerId: number;
  name: string;
  email: string;
  phone: string;
  address: string;
}